pyproject.toml
[tool.poetry]
name = "lms"
version = "0.3.0"
description = "New LMS for Insperia School with Soho.LMS"
authors = ["Sergey Natalenko <sergey.natalenko@mail.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "lms"}]

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.0.3"
SQLAlchemy = "^2.0.22"
pydantic = {extras = ["email"], version = "^2.4.2"}
orjson = "^3.9.9"
greenlet = "^3.0.0"
itsdangerous = "^2.1.2"
configargparse = "^1.7"
aiomisc-dependency = "^0.1.20"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
google-api-service-helper = "^0.1.1"
aiomisc = {extras = ["asgi", "cron", "rich", "uvicorn", "uvloop"], version = "^17.3.41"}
fastapi = "0.104.1"
alembic = "^1.12.0"
uvicorn = "^0.23.2"
asyncpg = "^0.30.0"
python-jose = "^3.3.0"
gunicorn = "^21.2.0"
wtforms = "^3.1.2"
wtforms-sqlalchemy = "^0.4.1"
aiocache = "^0.12.2"
sqladmin = {extras = ["full"], version = "^0.16.0"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
mypy = "^1.6.0"
types-python-jose = "^3.3.4.8"
pytest = "^7.4.2"
httpx = "^0.28.1"
pytest-asyncio = "^0.21.1"
pytest-deadfixtures = "^2.2.1"
polyfactory = "^2.9.0"
pytest-cov = "^4.1.0"
pytest-subtests = "^0.11.0"
types-passlib = "^1.7.7.13"
aiomisc-pytest = "^1.1.1"
ruff = "0.3.7"
factory-boy = "^3.3.0"
dirty-equals = "^0.7.1.post0"

[tool.poem-plugins.version]
provider = "git"
update_pyproject = true

[tool.poem-plugins.version.git]
format = "short"
prefix = "v"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ["py311"]

[tool.isort]
known_local_folder = ["lms", "tests"]
py_version = "311"
profile = "black"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
asyncio_mode = "auto"

[tool.ruff]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
]

line-length = 88
indent-width = 4

target-version = "py311"

[tool.ruff.lint]
select = [
    "BLE",
    "C90",
    "E",
    "F",
    "G",
    "I",
    "ICN",
    "ISC",
    "PLE",
    "Q",
    "RUF006",
    "RUF100",
    "T10",
    "T20",
    "TID",
    "UP",
    "W",
]
ignore = ["ISC001", "RUF006"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = false
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "aiocache.*",
    "aiomisc_dependency.*",
    "configargparse.*",
    "google_api_service_helper.*",
]
ignore_missing_imports = true
